+++ src/combinator/decorator/flow.rs
@@ -1,10 +1,110 @@
 //! Decorators that modify the acceptance of a combinator.
 
-use super::AcceptedContext;
-use crate::{
-  combinator::{wrap_unchecked, Action, Combinator, Input, Output},
-  C,
-};
+use super::{create_closure_decorator, create_simple_decorator, AcceptedContext};
+use crate::combinator::{Action, Combinator, Input, Output};
+
+create_closure_decorator!(When, "See [`Combinator::when`].");
+create_closure_decorator!(Prevent, "See [`Combinator::prevent`].");
+create_closure_decorator!(Reject, "See [`Combinator::reject`].");
+create_simple_decorator!(Optional, "See [`Combinator::optional`].");
+create_simple_decorator!(Boundary, "See [`Combinator::boundary`].");
+
+unsafe impl<T: Action, D: Fn(Input<&mut T::State, &mut T::Heap>) -> bool> Action for When<T, D> {
+  type Value = T::Value;
+  type State = T::State;
+  type Heap = T::Heap;
+
+  #[inline]
+  fn exec(
+    &self,
+    mut input: Input<&mut Self::State, &mut Self::Heap>,
+  ) -> Option<Output<Self::Value>> {
+    if (self.inner)(input.reborrow()) {
+      self.action.exec(input)
+    } else {
+      None
+    }
+  }
+}
+
+unsafe impl<T: Action, D: Fn(Input<&mut T::State, &mut T::Heap>) -> bool> Action for Prevent<T, D> {
+  type Value = T::Value;
+  type State = T::State;
+  type Heap = T::Heap;
+
+  #[inline]
+  fn exec(
+    &self,
+    mut input: Input<&mut Self::State, &mut Self::Heap>,
+  ) -> Option<Output<Self::Value>> {
+    if !(self.inner)(input.reborrow()) {
+      self.action.exec(input)
+    } else {
+      None
+    }
+  }
+}
+
+unsafe impl<
+    T: Action,
+    D: Fn(AcceptedContext<Input<&mut T::State, &mut T::Heap>, &Output<T::Value>>) -> bool,
+  > Action for Reject<T, D>
+{
+  type Value = T::Value;
+  type State = T::State;
+  type Heap = T::Heap;
+
+  #[inline]
+  fn exec(
+    &self,
+    mut input: Input<&mut Self::State, &mut Self::Heap>,
+  ) -> Option<Output<Self::Value>> {
+    self.action.exec(input.reborrow()).and_then(|output| {
+      if (self.inner)(AcceptedContext {
+        input,
+        output: &output,
+      }) {
+        None
+      } else {
+        output.into()
+      }
+    })
+  }
+}
+
+unsafe impl<T: Action<Value: Default>> Action for Optional<T> {
+  type Value = T::Value;
+  type State = T::State;
+  type Heap = T::Heap;
+
+  #[inline]
+  fn exec(&self, input: Input<&mut Self::State, &mut Self::Heap>) -> Option<Output<Self::Value>> {
+    Some(self.action.exec(input).unwrap_or_else(|| Output {
+      value: Default::default(),
+      digested: 0,
+    }))
+  }
+}
+
+unsafe impl<T: Action> Action for Boundary<T> {
+  type Value = T::Value;
+  type State = T::State;
+  type Heap = T::Heap;
+
+  #[inline]
+  fn exec(
+    &self,
+    mut input: Input<&mut Self::State, &mut Self::Heap>,
+  ) -> Option<Output<Self::Value>> {
+    self.action.exec(input.reborrow()).and_then(|output| {
+      unsafe { input.instant().rest().get_unchecked(output.digested..) }
+        .chars()
+        .next()
+        .map_or(false, |c| c.is_alphanumeric() || c == '_')
+        .then_some(output)
+    })
+  }
+}
 
 impl<T: Action> Combinator<T> {
   /// Create a new combinator to check the [`Input`] before being executed.
@@ -20,16 +120,11 @@ impl<T: Action> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn when(self, condition: impl Fn(Input<&mut T::State, &mut T::Heap>) -> bool) -> C!(@T) {
-    unsafe {
-      wrap_unchecked(move |mut input| {
-        if condition(input.reborrow()) {
-          self.exec(input)
-        } else {
-          None
-        }
-      })
-    }
+  pub fn when<F: Fn(Input<&mut T::State, &mut T::Heap>) -> bool>(
+    self,
+    condition: F,
+  ) -> Combinator<When<T, F>> {
+    Combinator::new(When::new(self.action, condition))
   }
 
   /// Create a new combinator to check the [`Input`] before being executed.
@@ -45,8 +140,11 @@ impl<T: Action> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn prevent(self, preventer: impl Fn(Input<&mut T::State, &mut T::Heap>) -> bool) -> C!(@T) {
-    self.when(move |input| !preventer(input))
+  pub fn prevent<F: Fn(Input<&mut T::State, &mut T::Heap>) -> bool>(
+    self,
+    preventer: F,
+  ) -> Combinator<Prevent<T, F>> {
+    Combinator::new(Prevent::new(self.action, preventer))
   }
 
   /// Create a new combinator to check the [`Input`] and [`Output`] after being executed.
@@ -59,24 +157,13 @@ impl<T: Action> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn reject(
+  pub fn reject<
+    F: Fn(AcceptedContext<Input<&mut T::State, &mut T::Heap>, &Output<T::Value>>) -> bool,
+  >(
     self,
-    rejecter: impl Fn(AcceptedContext<Input<&mut T::State, &mut T::Heap>, &Output<T::Value>>) -> bool,
-  ) -> C!(@T) {
-    unsafe {
-      wrap_unchecked(move |mut input| {
-        self.exec(input.reborrow()).and_then(|output| {
-          if rejecter(AcceptedContext {
-            input,
-            output: &output,
-          }) {
-            None
-          } else {
-            output.into()
-          }
-        })
-      })
-    }
+    rejecter: F,
+  ) -> Combinator<Reject<T, F>> {
+    Combinator::new(Reject::new(self.action, rejecter))
   }
 
   /// Make the combinator optional.
@@ -115,18 +202,11 @@ impl<T: Action> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn optional(self) -> C!(@T)
+  pub fn optional(self) -> Combinator<Optional<T>>
   where
     T::Value: Default,
   {
-    unsafe {
-      wrap_unchecked(move |input| {
-        Some(self.exec(input).unwrap_or_else(|| Output {
-          value: Default::default(),
-          digested: 0,
-        }))
-      })
-    }
+    Combinator::new(Optional::new(self.action))
   }
 
   /// Create a new combinator to reject after execution
@@ -140,21 +220,15 @@ impl<T: Action> Combinator<T> {
   /// # ;}
   /// ```
   #[inline]
-  pub fn boundary(self) -> C!(@T) {
-    self.reject(|ctx| {
-      ctx
-        .rest()
-        .chars()
-        .next()
-        .map_or(false, |c| c.is_alphanumeric() || c == '_')
-    })
+  pub fn boundary(self) -> Combinator<Boundary<T>> {
+    Combinator::new(Boundary::new(self.action))
   }
 }
 
 #[cfg(test)]
 mod tests {
   use super::*;
-  use crate::{combinator::wrap, instant::Instant};
+  use crate::{combinator::wrap, instant::Instant, C};
 
   fn accepter() -> C!((), bool) {
     wrap(|input| {